+++ lang/src/accounts/mod.rs
@@ -2,6 +2,7 @@
 
 pub mod account;
 pub mod account_info;
+pub mod account_loader;
 pub mod boxed;
 #[doc(hidden)]
 #[allow(deprecated)]
@@ -12,7 +13,6 @@ pub mod cpi_state;
 #[doc(hidden)]
 #[allow(deprecated)]
 pub mod loader;
-pub mod loader_account;
 pub mod program;
 #[doc(hidden)]
 #[allow(deprecated)]
